fix(#368 | Color Knob): remove color picker and split widgets in more files
