--- packages/widgetbook/lib/src/fields/color_field.dart
@@ -1,770 +1,0 @@
-import 'dart:math';
-
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-import 'field.dart';
-import 'field_codec.dart';
-import 'field_type.dart';
-
-enum ColorSpace {
-  rgba,
-  hex,
-  hsl,
-}
-
-/// [Field] that builds [ColorsFieldWidget] for [Color] values.
-///
-/// The [ColorField] uses the [ColorFieldCodec] to convert the [Color] value to a
-/// hex string and vice versa.
-///
-/// The [ColorField] uses the [ColorSpace] to determine which format the color is.
-class ColorField extends Field<Color> {
-  ColorField({
-    required super.name,
-    super.initialValue = defaultColor,
-    this.initialColorSpace = ColorSpace.hex,
-    super.onChanged,
-  }) : super(
-          type: FieldType.color,
-          codec: FieldCodec(
-            toParam: (color) => color.value.toRadixString(16),
-            toValue: (param) {
-              if (param == null) return null;
-              if (param == '0') return Colors.transparent;
-              if (param.length == 6) param = 'ff' + param;
-              return Color(
-                int.parse(
-                  param,
-                  radix: 16,
-                ),
-              );
-            },
-          ),
-        );
-
-  ColorSpace initialColorSpace;
-
-  static const defaultColor = Colors.white;
-
-  @override
-  Widget toWidget(BuildContext context, String group, Color? value) {
-    return ColorsFieldWidget(
-      name: name,
-      colorSpace: initialColorSpace,
-      value: value,
-      initialValue: initialValue,
-      defaultColor: defaultColor,
-      paramValue: codec.toParam(value ?? initialValue ?? defaultColor),
-      onChanged: (value) {
-        updateField(
-          context,
-          group,
-          codec.toValue(value as String?) ?? initialValue ?? defaultColor,
-        );
-      },
-    );
-  }
-}
-
-class ColorsFieldWidget extends StatefulWidget {
-  const ColorsFieldWidget({
-    required this.name,
-    required this.colorSpace,
-    required this.paramValue,
-    required this.initialValue,
-    required this.value,
-    required this.defaultColor,
-    required this.onChanged,
-    super.key,
-  });
-
-  final ColorSpace colorSpace;
-  final String name;
-  final Color? value;
-  final Color? initialValue;
-  final Color defaultColor;
-  final String paramValue;
-  final ValueChanged<dynamic> onChanged;
-
-  @override
-  State<ColorsFieldWidget> createState() => _ColorsFieldWidgetState();
-}
-
-class _ColorsFieldWidgetState extends State<ColorsFieldWidget> {
-  late ColorSpace initialColorSpace;
-  late dynamic colorValue;
-  bool shouldUpdateFromPicker = false;
-  final ColorsConverter converter = ColorsConverter();
-
-  @override
-  void initState() {
-    super.initState();
-    initialColorSpace = widget.colorSpace;
-    colorValue = converter.getValueByColorSpace(
-      colorSpace: initialColorSpace,
-      value: widget.paramValue,
-    );
-  }
-
-  void _onTapColorPicker(String color) {
-    widget.onChanged(
-      converter.convertColorValueToHex<String>(
-        colorSpace: ColorSpace.hex,
-        colorValues: color,
-      ),
-    );
-    setState(() {
-      colorValue = converter.getValueByColorSpace(
-        colorSpace: initialColorSpace,
-        value: color,
-      );
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        if (initialColorSpace == ColorSpace.rgba) ...[
-          RgbaColorTextFields(
-            colorValue: [...colorValue as List<String>],
-            shouldUpdateFromPicker: shouldUpdateFromPicker,
-            onChanged: (value) {
-              widget.onChanged(value);
-              setState(() {
-                colorValue = converter.getValueByColorSpace(
-                  colorSpace: initialColorSpace,
-                  value: widget.paramValue,
-                );
-              });
-            },
-            converter: converter,
-          )
-        ] else if (initialColorSpace == ColorSpace.hsl) ...[
-          HslColorTextFields(
-            colorValue: [...colorValue as List<String>],
-            paramValue: widget.paramValue,
-            shouldUpdateFromPicker: shouldUpdateFromPicker,
-            onChanged: (hexValue, value) {
-              if (hexValue != null) {
-                widget.onChanged(hexValue);
-              }
-              setState(() {
-                colorValue = [...value];
-              });
-            },
-            converter: converter,
-          )
-        ] else ...[
-          Row(
-            children: [
-              ColorTextField(
-                key: const Key('colorFieldHex'),
-                value: '$colorValue',
-                shouldUpdateFromPicker: shouldUpdateFromPicker,
-                maxLength: 8,
-                prefixIcon: const Icon(Icons.numbers),
-                inputFormatters: [
-                  FilteringTextInputFormatter.allow(
-                    RegExp(r'^[0-9a-fA-F]{0,8}$'),
-                    replacementString: '$colorValue',
-                  ),
-                ],
-                onChanged: (value) {
-                  widget.onChanged(
-                    converter.convertColorValueToHex<String>(
-                      colorSpace: initialColorSpace,
-                      colorValues: value,
-                    ),
-                  );
-                },
-              ),
-            ],
-          ),
-        ],
-        const SizedBox(
-          height: 16,
-        ),
-        Row(
-          children: [
-            MenuAnchor(
-              alignmentOffset: const Offset(-256, 0),
-              onClose: () {
-                setState(() {
-                  shouldUpdateFromPicker = false;
-                });
-              },
-              onOpen: () {
-                _onTapColorPicker(widget.paramValue);
-                setState(() {
-                  shouldUpdateFromPicker = true;
-                });
-              },
-              builder: (context, controller, child) {
-                return IconButton(
-                  key: const Key('colorFieldMenuButton'),
-                  onPressed: () {
-                    if (controller.isOpen) {
-                      controller.close();
-                    } else {
-                      controller.open();
-                    }
-                  },
-                  icon: Icon(
-                    Icons.square,
-                    color: widget.value,
-                  ),
-                );
-              },
-              menuChildren: [
-                Container(
-                  padding: const EdgeInsets.all(16),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      const Text(
-                        'Color Picker',
-                        style: TextStyle(
-                          fontSize: 16,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(
-                        height: 16,
-                      ),
-                      Container(
-                        height: 256,
-                        width: 256,
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(
-                            4,
-                          ),
-                          boxShadow: [
-                            BoxShadow(
-                              color: Colors.black.withOpacity(0.2),
-                              blurRadius: 4,
-                              offset: const Offset(0, 2),
-                            ),
-                          ],
-                        ),
-                        clipBehavior: Clip.hardEdge,
-                        child: GridView.builder(
-                          gridDelegate:
-                              const SliverGridDelegateWithFixedCrossAxisCount(
-                            crossAxisCount: 4,
-                            mainAxisExtent: 64,
-                          ),
-                          itemCount: 16,
-                          itemBuilder: (context, index) {
-                            return InkWell(
-                              key: Key('colorPickerItem$index'),
-                              onTap: () {
-                                _onTapColorPicker(
-                                  Colors.primaries[index].value
-                                      .toRadixString(16),
-                                );
-                              },
-                              child: Container(
-                                width: 64,
-                                height: 64,
-                                color: Colors.primaries[index],
-                                child: Icon(
-                                  Icons.check,
-                                  color: widget.value?.value ==
-                                          Colors.primaries[index].value
-                                      ? Colors.white
-                                      : Colors.transparent,
-                                  size: 32,
-                                ),
-                              ),
-                            );
-                          },
-                        ),
-                      ),
-                    ],
-                  ),
-                )
-              ],
-            ),
-            const SizedBox(
-              width: 8,
-            ),
-            Flexible(
-              child: DropdownButtonFormField<ColorSpace>(
-                value: initialColorSpace,
-                onChanged: (value) {
-                  setState(() {
-                    initialColorSpace = value ?? initialColorSpace;
-                    colorValue = converter.getValueByColorSpace(
-                      colorSpace: initialColorSpace,
-                      value: widget.paramValue,
-                    );
-                  });
-                },
-                items: ColorSpace.values
-                    .map(
-                      (value) => DropdownMenuItem(
-                        value: value,
-                        child: Text(value.name),
-                      ),
-                    )
-                    .toList(),
-              ),
-            ),
-          ],
-        ),
-      ],
-    );
-  }
-}
-
-class RgbaColorTextFields extends StatelessWidget {
-  const RgbaColorTextFields({
-    required this.colorValue,
-    required this.onChanged,
-    required this.converter,
-    required this.shouldUpdateFromPicker,
-    super.key,
-  });
-
-  final List<String> colorValue;
-  final ValueChanged<String> onChanged;
-  final ColorsConverter converter;
-  final bool shouldUpdateFromPicker;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      key: key,
-      children: [
-        ColorTextField(
-          key: const Key('colorFieldRgbaRed'),
-          value: colorValue[1],
-          shouldUpdateFromPicker: shouldUpdateFromPicker,
-          maxLength: 3,
-          inputFormatters: [
-            FilteringTextInputFormatter.digitsOnly,
-            FilteringTextInputFormatter.allow(
-              RegExp(r'^(?:1?[0-9]{1,2}|2[0-4][0-9]|25[0-5])$'),
-              replacementString: '${colorValue[1]}',
-            )
-          ],
-          labelText: 'R',
-          onChanged: (value) => onChanged(
-            converter.convertColorValueToHex<List<String>>(
-              colorSpace: ColorSpace.rgba,
-              colorValues: [
-                '${colorValue[0]}',
-                value,
-                '${colorValue[2]}',
-                '${colorValue[3]}',
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(
-          width: 8,
-        ),
-        ColorTextField(
-          key: const Key('colorFieldRgbaGreen'),
-          value: '${colorValue[2]}',
-          shouldUpdateFromPicker: shouldUpdateFromPicker,
-          maxLength: 3,
-          inputFormatters: [
-            FilteringTextInputFormatter.digitsOnly,
-            FilteringTextInputFormatter.allow(
-              RegExp(r'^(?:1?[0-9]{1,2}|2[0-4][0-9]|25[0-5])$'),
-              replacementString: '${colorValue[2]}',
-            )
-          ],
-          labelText: 'G',
-          onChanged: (value) => onChanged(
-            converter.convertColorValueToHex<List<String>>(
-              colorSpace: ColorSpace.rgba,
-              colorValues: [
-                '${colorValue[0]}',
-                '${colorValue[1]}',
-                value,
-                '${colorValue[3]}',
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(
-          width: 8,
-        ),
-        ColorTextField(
-          key: const Key('colorFieldRgbaBlue'),
-          value: '${colorValue[3]}',
-          shouldUpdateFromPicker: shouldUpdateFromPicker,
-          maxLength: 3,
-          inputFormatters: [
-            FilteringTextInputFormatter.digitsOnly,
-            FilteringTextInputFormatter.allow(
-              RegExp(r'^(?:1?[0-9]{1,2}|2[0-4][0-9]|25[0-5])$'),
-              replacementString: '${colorValue[3]}',
-            )
-          ],
-          labelText: 'B',
-          onChanged: (value) => onChanged(
-            converter.convertColorValueToHex<List<String>>(
-              colorSpace: ColorSpace.rgba,
-              colorValues: [
-                '${colorValue[0]}',
-                '${colorValue[1]}',
-                '${colorValue[2]}',
-                value,
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(
-          width: 8,
-        ),
-        ColorTextField(
-          key: const Key('colorFieldRgbaAlpha'),
-          value: '${colorValue[0]}',
-          shouldUpdateFromPicker: shouldUpdateFromPicker,
-          maxLength: 3,
-          inputFormatters: [
-            FilteringTextInputFormatter.digitsOnly,
-            FilteringTextInputFormatter.allow(
-              RegExp(r'^(?:1?[0-9]{1,2}|2[0-4][0-9]|25[0-5])$'),
-              replacementString: '${colorValue[0]}',
-            )
-          ],
-          labelText: 'A',
-          onChanged: (value) => onChanged(
-            converter.convertColorValueToHex<List<String>>(
-              colorSpace: ColorSpace.rgba,
-              colorValues: [
-                value,
-                '${colorValue[1]}',
-                '${colorValue[2]}',
-                '${colorValue[3]}',
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class HslColorTextFields extends StatelessWidget {
-  const HslColorTextFields({
-    required this.colorValue,
-    required this.onChanged,
-    required this.paramValue,
-    required this.converter,
-    required this.shouldUpdateFromPicker,
-    super.key,
-  });
-
-  final List<String> colorValue;
-  final void Function(String? hexValue, List<String> newValues) onChanged;
-  final bool shouldUpdateFromPicker;
-  final String paramValue;
-  final ColorsConverter converter;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      key: key,
-      children: [
-        ColorTextField(
-          key: const Key('colorFieldHslHue'),
-          value: '${colorValue[0]}',
-          shouldUpdateFromPicker: shouldUpdateFromPicker,
-          maxLength: 3,
-          inputFormatters: [
-            FilteringTextInputFormatter.digitsOnly,
-            FilteringTextInputFormatter.allow(
-              RegExp(r'^(?:[0-9]\d?|[12]\d{2}|3[0-5]\d)$'),
-              replacementString: '${colorValue[0]}',
-            )
-          ],
-          labelText: 'H',
-          onChanged: (value) => checkHueAndSaturation([
-            value,
-            '${colorValue[1]}',
-            '${colorValue[2]}',
-          ]),
-        ),
-        const SizedBox(
-          width: 8,
-        ),
-        ColorTextField(
-          key: const Key('colorFieldHslSaturation'),
-          value: '${colorValue[1]}',
-          shouldUpdateFromPicker: shouldUpdateFromPicker,
-          maxLength: 3,
-          inputFormatters: [
-            FilteringTextInputFormatter.digitsOnly,
-            FilteringTextInputFormatter.allow(
-              RegExp(r'^(0|[1-9][0-9]?|100)$'),
-              replacementString: '${colorValue[1]}',
-            )
-          ],
-          labelText: 'S',
-          suffixText: '%',
-          onChanged: (value) => checkHueAndSaturation([
-            '${colorValue[0]}',
-            value,
-            '${colorValue[2]}',
-          ]),
-        ),
-        const SizedBox(
-          width: 8,
-        ),
-        ColorTextField(
-          key: const Key('colorFieldHslLightness'),
-          value: '${colorValue[2]}',
-          shouldUpdateFromPicker: shouldUpdateFromPicker,
-          maxLength: 3,
-          inputFormatters: [
-            FilteringTextInputFormatter.digitsOnly,
-            FilteringTextInputFormatter.allow(
-              RegExp(r'^(0|[1-9][0-9]?|100)$'),
-              replacementString: '${colorValue[2]}',
-            )
-          ],
-          labelText: 'L',
-          suffixText: '%',
-          onChanged: (value) => checkLightness([
-            '${colorValue[0]}',
-            '${colorValue[1]}',
-            value,
-          ]),
-        ),
-      ],
-    );
-  }
-
-  void checkHueAndSaturation(List<String> updatedValues) {
-    final initialValue = converter.convertColorValueToHex<List<String>>(
-      colorSpace: ColorSpace.hsl,
-      colorValues: converter.getValueByColorSpace(
-        colorSpace: ColorSpace.hsl,
-        value: paramValue,
-      ) as List<String>,
-    );
-    final updateValue = converter.convertColorValueToHex<List<String>>(
-      colorSpace: ColorSpace.hsl,
-      colorValues: updatedValues,
-    );
-    if (initialValue != updateValue) {
-      onChanged(updateValue, updatedValues);
-    } else {
-      onChanged(null, updatedValues);
-    }
-  }
-
-  void checkLightness(List<String> updatedValues) {
-    final initialValue = converter.convertColorValueToHex<List<String>>(
-      colorSpace: ColorSpace.hsl,
-      colorValues: converter.getValueByColorSpace(
-        colorSpace: ColorSpace.hsl,
-        value: paramValue,
-      ) as List<String>,
-    );
-    final updateValue = converter.convertColorValueToHex<List<String>>(
-      colorSpace: ColorSpace.hsl,
-      colorValues: updatedValues,
-    );
-    if (initialValue != updateValue) {
-      onChanged(updateValue, updatedValues);
-    }
-  }
-}
-
-class ColorTextField extends StatelessWidget {
-  const ColorTextField({
-    required this.value,
-    required this.maxLength,
-    required this.onChanged,
-    required this.shouldUpdateFromPicker,
-    this.inputFormatters,
-    this.prefixIcon,
-    this.suffixText,
-    this.labelText,
-    super.key,
-  });
-
-  final String value;
-  final int maxLength;
-  final ValueChanged<String> onChanged;
-  final List<TextInputFormatter>? inputFormatters;
-  final bool shouldUpdateFromPicker;
-  final Widget? prefixIcon;
-  final String? suffixText;
-  final String? labelText;
-
-  @override
-  Widget build(BuildContext context) {
-    return Flexible(
-      child: TextFormField(
-        key: shouldUpdateFromPicker ? Key('colorFieldPicker$value') : key,
-        initialValue: value,
-        inputFormatters: inputFormatters,
-        maxLength: maxLength,
-        decoration: InputDecoration(
-          labelText: labelText,
-          counterText: '',
-          prefixIcon: prefixIcon,
-          suffixText: suffixText,
-        ),
-        onChanged: onChanged,
-      ),
-    );
-  }
-}
-
-class ColorsConverter {
-  String convertColorValueToHex<T>({
-    required ColorSpace colorSpace,
-    required T colorValues,
-  }) {
-    switch (colorSpace) {
-      case ColorSpace.hex:
-        return colorValues as String;
-      case ColorSpace.rgba:
-        return rgbaToHex(colorValues as List<String>);
-      default:
-        return hslToHex(colorValues as List<String>);
-    }
-  }
-
-  String rgbaToHex(List<String> rgba) {
-    int alpha = int.tryParse(rgba[0]) ?? 0;
-    int red = int.tryParse(rgba[1]) ?? 0;
-    int green = int.tryParse(rgba[2]) ?? 0;
-    int blue = int.tryParse(rgba[3]) ?? 0;
-    return '${alpha.toRadixString(16).padLeft(2, '0')}${red.toRadixString(16).padLeft(2, '0')}${green.toRadixString(16).padLeft(2, '0')}${blue.toRadixString(16).padLeft(2, '0')}';
-  }
-
-  String hslToHex(List<String> hsl) {
-    double h = double.tryParse(hsl[0]) ?? 0;
-    double s = double.tryParse(hsl[1]) ?? 0;
-    double l = double.tryParse(hsl[2]) ?? 0;
-
-    h /= 360;
-    s /= 100;
-    l /= 100;
-
-    double r = 0;
-    double g = 0;
-    double b = 0;
-    if (s == 0) {
-      r = g = b = l;
-    } else {
-      double q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-      double p = 2 * l - q;
-      r = hueToRgb(p, q, h + 1 / 3);
-      g = hueToRgb(p, q, h);
-      b = hueToRgb(p, q, h - 1 / 3);
-    }
-
-    return rgbaToHex([
-      '255',
-      '${convertToRgbValue(r)}',
-      '${convertToRgbValue(g)}',
-      '${convertToRgbValue(b)}'
-    ]);
-  }
-
-  // To maximize precision when converting from hsl to rgb,
-  // we need to get the minimum value between the calculated value and 255
-  int convertToRgbValue(double value) {
-    return min((value * 256).floor(), 255);
-  }
-
-  // https://stackoverflow.com/questions/2353211/hsl-to-rgb-color-conversion
-  double hueToRgb(double p, double q, double t) {
-    if (t < 0) t += 1;
-    if (t > 1) t -= 1;
-    if (t < 1 / 6) return p + (q - p) * 6 * t;
-    if (t < 1 / 2) return q;
-    if (t < 2 / 3) return p + (q - p) * (2 / 3 - t) * 6;
-    return p;
-  }
-
-  dynamic getValueByColorSpace({
-    required ColorSpace colorSpace,
-    required String value,
-  }) {
-    switch (colorSpace) {
-      case ColorSpace.hex:
-        return value;
-      case ColorSpace.rgba:
-        return hexToRgba(value);
-      default:
-        return hexToHsl(value);
-    }
-  }
-
-  List<String> hexToRgba(String hex) {
-    // Add the alpha value if it's not there
-    if (hex.length == 1) {
-      hex = ''.padLeft(8, hex);
-    }
-    // Add the alpha value if it's not there
-    if (hex.length == 6) {
-      hex = 'ff' + hex;
-    }
-    // Get the int value of the hex string
-    int intColor = int.tryParse('0x$hex') ?? 0;
-    int alpha = (intColor >> 24) & 0xFF;
-    int red = (intColor >> 16) & 0xFF;
-    int green = (intColor >> 8) & 0xFF;
-    int blue = intColor & 0xFF;
-
-    return ['$alpha', '$red', '$green', '$blue'];
-  }
-
-  List<String> hexToHsl(String hex) {
-    // Start from rgba values because it's easier to convert to hsl
-    List<int> rgbaValues =
-        hexToRgba(hex).map((e) => int.tryParse(e) ?? 0).toList();
-
-    // get all the values and divide them by 255 to get a value between 0 and 1
-    double r = rgbaValues[1] / 255;
-    double g = rgbaValues[2] / 255;
-    double b = rgbaValues[3] / 255;
-
-    // get the max and min value of the rgb values
-    double max =
-        [r, g, b].reduce((value, element) => value > element ? value : element);
-    double min =
-        [r, g, b].reduce((value, element) => value < element ? value : element);
-
-    double h = 0;
-    double s = 0;
-    double l = (max + min) / 2;
-    if (max == min) {
-      h = s = 0;
-    } else {
-      // get the difference between the max and min value
-      double difference = max - min;
-      // calculate the saturation
-      s = l > 0.5 ? difference / (2 - max - min) : difference / (max + min);
-      // calculate the hue
-      if (max == r) {
-        h = (g - b) / difference + (g < b ? 6 : 0);
-      } else if (max == g) {
-        h = (b - r) / difference + 2;
-      } else if (max == b) {
-        h = (r - g) / difference + 4;
-      }
-      h /= 6;
-    }
-
-    return [
-      '${(h * 360).toInt()}',
-      '${(s * 100).toInt()}',
-      '${(l * 100).toInt()}'
-    ];
-  }
-}
