+++ packages/widgetbook/test/src/fields/color_field_test.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:widgetbook/src/fields/color_field/hsl_color_text_fields.dart';
+import 'package:widgetbook/src/fields/color_field/rgba_color_text_fields.dart';
 import 'package:widgetbook/widgetbook.dart';
 
 import '../../helper/helper.dart';
